requirements.txt
altair==5.5.0
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
bokeh==3.7.2
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.2.0
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
debugpy==1.8.14
decorator==5.2.1
distributed==2025.4.1
executing==2.2.0
findspark==2.0.1
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
importlib_metadata==8.7.0
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
locket==1.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
msgpack==1.1.0
narwhals==1.39.0
nest-asyncio==1.6.0
numpy==1.26.4
packaging==24.2
pandas==2.2.3
parso==0.8.4
partd==1.4.2
pillow==11.2.1
platformdirs==4.3.7
plotly==6.0.1
prompt_toolkit==3.0.51
protobuf==6.30.2
psutil==7.0.0
pure_eval==0.2.3
py4j==0.10.9.7
pyarrow==20.0.0
pydeck==0.9.1
Pygments==2.19.1
pyspark==3.5.5
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==310
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
six==1.17.0
smmap==5.0.2
sortedcontainers==2.4.0
soupsieve==2.7
stack-data==0.6.3
streamlit==1.45.1
tblib==3.1.0
tenacity==9.1.2
toml==0.10.2
toolz==1.0.0
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
wcwidth==0.2.13
xyzservices==2025.4.0
zict==3.0.0
zipp==3.21.0
